Refactored Footer
