+++ src/styles/sharedStyledComponents.tsx
@@ -1,0 +1,6 @@
+import { Divider } from "@blueprintjs/core";
+import styled from "styled-components";
+
+export const StyledDivider = styled(Divider)`
+    width: 100%;
+`
